Move error classes away from Runner
As the runner itself is becoming bigger, it's hard to navigate it with
those classes on the top.
